[ANDROSDK-1966] move fields and service to network layer
