+++ core/src/main/java/org/hisp/dhis/android/network/customIcon/CustomIconFields.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2004-2023, University of Oslo
+ *  Copyright (c) 2004-2025, University of Oslo
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -25,7 +25,7 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.icon.internal
+package org.hisp.dhis.android.network.customIcon
 
 import org.hisp.dhis.android.core.icon.CustomIcon
 import org.hisp.dhis.android.network.common.fields.BaseFields
